add CS1 Game Engine
The CS1 Game Engine is built on Three.js and A-Frame and intends to provide JavaScript developers with with a fast track to high performance 3D VR multiplayer games installable on phones, tablets, laptops, and desktops as Progressive Web Applications (PWAs). The resultant games are compatible with many popular VR HMD and controller systems such as the Oculus Quest.